DR 4.1
Added the n_value_calculator inline with Theory Prep. Select in the settings.yaml to choose if it is used or not.
Changed Peaks85 Converter to allow for extraction of multiple charge states of a single peptide. The range of charge states can be changed in settings in guide_settings.yaml.
Added correct calculation of n_isos in Peaks85 and now use them in Extractor.
Added fixes for windows staying open on a Linux machine
Updated aa_labeling_sites to include amino acids that weren't present but were present in DR3.35.