error check
added safety measure so that wrong peaks input headers don't error out.  Also added the make_folder function to delete the graph folder when going through rate calculation so old graphs are not preserved to cause issue.  not done to the main folder due to the possibility of starting partway through the worklist and possibility of unrelated files and subfolders.