statistical change
after discussion with a stats professor and looking at how to perform confidence interval for the curve_fit using stackoverflow and a similar site for stats, it is determined that the sqrt of the diagonal of the covariance matrix is the standard error not standard deviation.  In accordance with our discussion with the stats professor we have adjusted the confidence interval equation (and the header of the output file) to be appropriate.  This will increase confidence interval sizes