Chromatography division (#7)
* Updated with current chromatography division

Added files to github that has all chromatogrpahy division information.

* Visualizer and add Settings

Added visualizer to GitHub, Added more settings for what EIC should be used to divide the chromatography and how many peaks is the most used in the scoring algorithm.

* gitignore

* Added Brad's Protein Changes

Includes multiprocessing in FN and RC calculators, adding a column to the Labeling Table for Biological Replicates so I can use it for lipid de novo n-value comparisons, making the chromatography division be multiprocessed and memory efficient, and other cleaning up of the code.

* Relatively Final Lipid DeuteRater

Final changes to DR to make the chromatography division incorporated fully into deuterater, having Lipid Guide File settings available in the GUI, streamlining the NValueCalculator, and adding labeled progress bars.

* Removing Unused Files

Removing files that are not being used in DeuteRater that were saved to note what old versions looked like in-between commits.

* Update visualizer.py

Updated Visualizer to handle any number of neutromer peaks, the new chromatography division format, and lipids that were never extracted in a specific file.

* Abundance Fraction New Fix

Changed afn to abund_fn to clarify what the a in afn was. Made the abundance fraction new have options for how it is calculated (M0, most intense, median). Added Debug mode to Fraction New Calculator.

* Fixed Highest AFN

* Changed Lipid Converter Template

* Error Checking in Division Writing

* Fragments in PCDL Converter

* Added Debugging to Rate Calculator

* Changed How Baseline Calculated

* Cleanup

* Manual Merge

Brad and Chad worked together to merge the py 4.0 and chromatogrpahy_division branches. This is the final result. :)